requirements.txt
Django==1.10
django-filter==0.14.0
djangorestframework==3.4.6
Markdown==2.6.6
Pillow==3.3.1
pytest==3.0.1
requests==2.11.1
